+++ shuup_tests/admin/test_product_list_view.py
@@ -71,6 +71,7 @@ def test_list_view_with_multiple_suppliers(rf, admin_user):
         assert 200 <= response.status_code < 300
 
         data = json.loads(response.content.decode("utf-8"))
+        assert len(data["items"]) == 2
         product_data = [item for item in data["items"] if item["_id"] == shop_product.pk][0]
         assert product_data["primary_category"] == factories.get_default_category().name
         assert product_data["categories"] == factories.get_default_category().name
@@ -79,3 +80,16 @@ def test_list_view_with_multiple_suppliers(rf, admin_user):
 
         product_data2 = [item for item in data["items"] if item["_id"] == shop_product2.pk][0]
         assert product_data2["suppliers"] == factories.get_default_supplier().name
+
+    with override_settings(
+        SHUUP_ADMIN_SUPPLIER_PROVIDER_SPEC="shuup.testing.supplier_provider.FirstSupplierProvider"
+    ):
+        view = load("shuup.admin.modules.products.views:ProductListView").as_view()
+        request = apply_request_middleware(rf.get("/", {
+            "jq": json.dumps({"perPage": 100, "page": 1})
+        }), user=admin_user)
+        response = view(request)
+        assert 200 <= response.status_code < 300
+
+        data = json.loads(response.content.decode("utf-8"))
+        assert len(data["items"]) == 1