+++ shuup_tests/admin/test_variation_templates.py
@@ -18,18 +18,18 @@ def test_variation_template_creation(rf):
     parent = create_product(printable_gibberish())
     user = create_random_user(is_superuser = True, is_staff = True)
     request = apply_request_middleware(rf.get("/"), user=user, shop=shop)
-    form = VariationVariablesDataForm()
+    form = VariationVariablesDataForm(parent_product=parent, request=request)
     form.request = request
     form.cleaned_data = {'data' : "{}"}
     form.save()
     assert form.get_variation_templates() == []
-    form.cleaned_data = {'data': '{"variable_values": []}', 'activate_template': False, 'template_name': 'Test'}
+    form.cleaned_data = {'data': '{"variable_values": []}', 'template_name': 'Test'}
     form.save()
     var_template = form.get_variation_templates()[0]
     assert len(form.get_variation_templates()) == 1
     template_identifier = var_template.get('identifier')
-    template_data = {'data': '{"variable_values":' + str(var_data) + '}', 'activate_template': False, 'template_name': 'Test'}
-    form.cleaned_data = {'data': json.dumps({"variable_values": var_data, 'template_identifier' : template_identifier}), 'activate_template': False, 'template_name': '' }
+    template_data = {'data': '{"variable_values":' + str(var_data) + '}', 'template_name': 'Test'}
+    form.cleaned_data = {'data': json.dumps({"variable_values": var_data, 'template_identifier' : template_identifier}), 'template_name': '' }
     form.save()
     assert len(form.get_variation_templates()) == 1
     var_template = form.get_variation_templates()[0]
@@ -44,7 +44,11 @@ def test_variation_activation(rf):
     form = VariationVariablesDataForm(parent_product=parent, request=request)
     assert len(ProductVariationVariable.objects.all()) == 0  # Assert no active Variations are present
     assert form.get_variation_templates() == []  # Assert no templates exist (yet)
-    var_data_dict = {'data': json.dumps({"variable_values": var_data}), 'activate_template': True, 'template_name': ''}  # Base skeleton
+
+    var_data_dict = {
+        'data': json.dumps({"variable_values": var_data}),
+        'template_name': ''
+    }  # Base skeleton
     form.cleaned_data = var_data_dict
     form.is_valid()
     form.save()
@@ -52,8 +56,18 @@ def test_variation_activation(rf):
     assert len(ProductVariationVariable.objects.filter(product=parent)) == 1 # Size
     assert len(ProductVariationVariableValue.objects.all()) == 2  # X,S
 
-    var_data_one_value = [{"pk":"$0.3","identifier":"","names":{"en":"Size"},
-    "values":[{"pk":"$0.8","identifier":"","texts":{"en":"X"}}]}] #  Delete one value
+    var_data_one_value = [{
+        "pk": "$0.3",
+        "identifier": "",
+        "names": {"en": "Size"},
+        "values": [{
+            "pk": "$0.8",
+            "identifier": "",
+            "texts": {
+                "en": "X"
+            }
+        }]
+    }] #  Delete one value
     var_data_dict['data'] = json.dumps({"variable_values": var_data_one_value})
     form.cleaned_data = var_data_dict
     form.save()
@@ -88,13 +102,3 @@ def test_variation_activation(rf):
     assert len(ProductVariationVariableValue.objects.all()) == 1
     var_template_data = form.get_variation_templates()[0].get("data")
     assert var_template_data == var_data_one_value
-
-    var_data_dict['data'] = two_values_dict
-    var_data_dict['activate_template'] = False
-    form.cleaned_data = var_data_dict
-    form.save()
-
-    assert len(ProductVariationVariable.objects.filter(product=parent)) == 1
-    assert len(ProductVariationVariableValue.objects.all()) == 1  # Assert newest template data hasn't been activated since activated_template = False
-    var_template_data = form.get_variation_templates()[0].get("data")
-    assert var_template_data == var_data