+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 List all changes after the last release here (newer on top). Each change on a separate bullet point line.
 
 
+- Admin: slight improvement to product variation edit. Apply variation data everytime
+  for the current variation parent no matter whether using template or not.
+
+
 ## [2.1.5] - 2020-09-08
 
 ### Fixed